+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- \[Rust\]: Add support for strings in unions.
+- \[Rust\]: Add support for strings in unions
+- \[Rust\]: Add support for structs in unions
 
 ## [0.3.0] - 2022-02-06
 ### Added